+++ src/games/mod.rs
@@ -1,4 +1,0 @@
-pub mod snake;
-pub mod snake_board;
-pub mod drop;
-pub mod idle;